+++ TomasGreen.Model/Models/CompanySection.cs
@@ -7,7 +7,7 @@ namespace TomasGreen.Model.Models
 {
     public class CompanySection :BaseEntity
     {
-        public Int64 CompanyID { get; set; }
+        public int CompanyID { get; set; }
         [Required(ErrorMessage = "Name is required")]
         [StringLength(255, MinimumLength = 3, ErrorMessage = "Invalid string lenght")]
         public string Name { get; set; }
