+++ TomasGreen.Model/Models/DicingPlanDetail.cs
@@ -7,15 +7,15 @@ namespace TomasGreen.Model.Models
 {
     public class DicingPlanDetail :BaseEntity
     {
-        public Int64 DicingPlanID { get; set; }
+        public int DicingPlanID { get; set; }
         //Some dicing details
         //Out
         [Display(Name = "Warehouse")]
         [Required(ErrorMessage = "Please choose a warehouse.")]
-        public Int64 WarehouseID { get; set; }
+        public int WarehouseID { get; set; }
         [Display(Name = "Article")]
         [Required(ErrorMessage = "Please choose an article.")]
-        public Int64 ArticleID { get; set; }
+        public int ArticleID { get; set; }
         public int QtyPackages { get; set; }
         public decimal QtyExtra { get; set; }
         public decimal TotalWeight { get; set; }
