+++ TomasGreen.Model/Models/PackingPlanMix.cs
@@ -7,19 +7,19 @@ namespace TomasGreen.Model.Models
 {
     public class PackingPlanMix : BaseEntity
     {
-        public Int64 PackingPlanID { get; set; }
-        public Int64 PackagingMaterialPackageID { get; set; }
+        public int PackingPlanID { get; set; }
+        public int PackagingMaterialPackageID { get; set; }
         public int Packages { get; set; }
-        public Int64 PackagingMaterialBagID { get; set; }
+        public int PackagingMaterialBagID { get; set; }
         public int Bags { get; set; }
 
         //New article
         [Display(Name = "Warehouse")]
         [Required(ErrorMessage = "Please choose a warehouse.")]
-        public Int64 ToWarehouseID { get; set; }
+        public int ToWarehouseID { get; set; }
         [Display(Name = "Article")]
         [Required(ErrorMessage = "Please choose an article.")]
-        public Int64 NewArticleID { get; set; }
+        public int NewArticleID { get; set; }
         public int NewQtyPackages { get; set; }
         public decimal NewQtyExtra { get; set; }
         public decimal NewTotalWeight { get; set; }
