+++ TomasGreen.Model/Models/RoastingPlan.cs
@@ -12,19 +12,19 @@ namespace TomasGreen.Model.Models
         public DateTime Date { get; set; }
         [Display(Name = "Manager")]
         [Required(ErrorMessage = "Please choose a manager.")]
-        public Int64 ManagerID { get; set; }
+        public int ManagerID { get; set; }
         [Display(Name = "Company")]
         [Required(ErrorMessage = "Please choose a company.")]
-        public Int64 CompanyID { get; set; }
+        public int CompanyID { get; set; }
 
 
         //In
         [Display(Name = "Warehouse")]
         [Required(ErrorMessage = "Please choose a warehouse.")]
-        public Int64 FromWarehouseID { get; set; }
+        public int FromWarehouseID { get; set; }
         [Display(Name = "Article")]
         [Required(ErrorMessage = "Please choose an article.")]
-        public Int64 ArticleID { get; set; }
+        public int ArticleID { get; set; }
         public int QtyPackages { get; set; }
         public decimal QtyExtra { get; set; }
         public decimal TotalWeight { get; set; }
@@ -33,18 +33,18 @@ namespace TomasGreen.Model.Models
         public decimal WeightChange { get; set; }
         public decimal Salt { get; set; }
         public decimal SaltLimit { get; set; }// 2% of total weight
-        public Int64 PackagingMaterialPackageID { get; set; }
+        public int PackagingMaterialPackageID { get; set; }
         public int Packages{ get; set; }
-        public Int64 PackagingMaterialBagID { get; set; }
+        public int PackagingMaterialBagID { get; set; }
         public int Bags { get; set; }
 
         //New article
         [Display(Name = "Warehouse")]
         [Required(ErrorMessage = "Please choose a warehouse.")]
-        public Int64 ToWarehouseID { get; set; }
+        public int ToWarehouseID { get; set; }
         [Display(Name = "Article")]
         [Required(ErrorMessage = "Please choose an article.")]
-        public Int64 NewArticleID { get; set; }
+        public int NewArticleID { get; set; }
         public int NewQtyPackages { get; set; }
         public decimal NewQtyExtra { get; set; }
         public decimal NewTotalWeight { get; set; }
