+++ TomasGreen.Web/Areas/Import/ViewModels/SavePurchasedArticleWarehouseViewModel.cs
@@ -8,7 +8,7 @@ namespace TomasGreen.Web.Areas.Import.ViewModels
 {
     public class SavePurchasedArticleWarehouseViewModel
     {
-        public Int64 ID { get; set; }
+        public int ID { get; set; }
         //public List<Article> Articles { get; set; }
         //public List<Company> Companies { get; set; }
         public PurchasedArticle PurchasedArticle { get; set; }
