+++ TomasGreen.Web/Balances/ArticleInOut.cs
@@ -7,9 +7,9 @@ namespace TomasGreen.Web.Balances
 {
     public class ArticleInOut
     {
-        public Int64 ArticleID { get; set; }
-        public Int64 WarehouseID { get; set; }
-        public Int64 CompanyID { get; set; }
+        public int ArticleID { get; set; }
+        public int WarehouseID { get; set; }
+        public int CompanyID { get; set; }
         public int QtyPackagesIn { get; set; }
         public decimal QtyExtraIn { get; set; }
